Better code

+++ PAC/common/PAC_dev.cpp
@@ -1251,11 +1251,10 @@ device* device_manager::get_device( const char* dev_name )
     return get_stub_device();
     }
 //-----------------------------------------------------------------------------
-u_int device_manager::get_device_count() const
-
-{
+size_t device_manager::get_device_count() const
+    {
     return project_devices.size();
-}
+    }
 //-----------------------------------------------------------------------------
 void device_manager::print() const
     {
